Create a deque data type and update makefile.
It isn't a true deque because you can peak and pop from the middle and a
random index. The deque is for the Growing Tree algorithm. This
algorithm can be customized to choose cells from a list by newest,
oldest, middle, and random. A circular array was chosen to store this.
Removing the front and back makes a linked list preferable over an array,
since removing from the front and back of a linked list and removing from
the back of an array is O(1), but removing from the front of an array is
O(n). However, the middle and random require accessing arbitrary indexes,
making an array preferable over a linked list since accessing arbitrary
indexes in a linked list is O(n) and accessing arbitrary indexes in an
array is O(1). However, if we use a circular array, we can access in O(1),
remove the front and back at O(1), and mostly insert in O(1) except when
we need to increase the size of the array, which is O(n). Removing from
the middle and random is still O(n).